moved offline abstract part out of field monitoring
