+++ src/etools/applications/offline/errors.py
