+++ src/etools/applications/offline/metadata.py
