+++ src/etools/applications/offline/tests/test_contact_book_example.py
@@ -1,0 +1,63 @@
+import json
+import os
+
+from django.test import TestCase
+
+from etools.applications.offline.errors import ValidationError
+from etools.applications.offline.tests.contact_book import contact_book
+
+
+class ContactBookExampleTestCase(TestCase):
+    @classmethod
+    def setUpTestData(cls):
+        super().setUpTestData()
+
+    def test_serialized_structure(self):
+        with open(os.path.join(os.path.dirname(os.path.abspath(__file__)), 'contact_book.json'), 'r') as example:
+            self.assertDictEqual(json.load(example), contact_book.to_dict())
+
+    def test_form_validation(self):
+        contact_book.validate({
+            'name': 'test book',
+            'users': [
+                {
+                    'full_name': 'John Doe',
+                    'phones': [
+                        {'number': '1234567', 'type': 'mobile'},
+                        {'number': '2222222', 'type': 'work'},
+                        {'number': '2222242'},
+                    ],
+                    'groups': ['test']  # todo: how resolve options? metadata? if remote choices? hidden validator?
+                },
+            ]
+        })
+
+    def test_users_not_required(self):
+        contact_book.validate({'name': 'test'})
+
+    def test_phones_required(self):
+        value = {'name': 'test', 'users': [{'full_name': 'John Doe'}]}
+        with self.assertRaises(ValidationError) as err:
+            contact_book.validate(value)
+        self.assertDictEqual({'users': [{'phones': ['This field is required']}]}, err.exception.detail)
+
+        value['users'][0]['phones'] = [{'number': '1234567'}]
+        contact_book.validate(value)
+
+    def test_name_required(self):
+        value = {'name': None}
+        with self.assertRaises(ValidationError) as err:
+            contact_book.validate(value)
+        self.assertDictEqual({'name': ['This field is required']}, err.exception.detail)
+
+        value['name'] = 'test'
+        contact_book.validate(value)
+
+    def test_number_regex(self):
+        value = {'name': 'test', 'users': [{'full_name': 'John Doe', 'phones': [{'number': '123456'}]}]}
+        with self.assertRaises(ValidationError) as err:
+            contact_book.validate(value)
+        self.assertDictEqual({'users': [{'phones': [{'number': [{'Invalid value: 123456'}]}]}]}, err.exception.detail)
+
+        value['users'][0]['phones'][0]['number'] = '1234567'
+        contact_book.validate(value)
