+++ src/etools/applications/offline/tests/test_validations.py
@@ -1,0 +1,1 @@
+# todo
