+++ src/etools/applications/offline/validations/base.py
