+++ src/etools/applications/offline/validations/choices.py
@@ -1,5 +1,5 @@
-from etools.applications.field_monitoring.data_collection.offline.validations.base import BaseValidation
-from etools.applications.field_monitoring.data_collection.offline.validations.errors import BadValueError
+from etools.applications.offline.errors import BadValueError
+from etools.applications.offline.validations.base import BaseValidation
 
 
 class Choice:
