+++ src/etools/applications/offline/validations/text.py
@@ -1,10 +1,7 @@
 import re
 
-from etools.applications.field_monitoring.data_collection.offline.validations.base import BaseValidation
-from etools.applications.field_monitoring.data_collection.offline.validations.errors import (
-    BadValueError,
-    ValueTypeMismatch,
-)
+from etools.applications.offline.errors import BadValueError, ValueTypeMismatch
+from etools.applications.offline.validations.base import BaseValidation
 
 
 class TextValidation(BaseValidation):
